ocr-server/requirements.txt
rapidocr-onnxruntime
fastapi
uvicorn
python-multipart
Pillow
numpy
